requirements.txt
attrs==21.4.0
iniconfig==1.1.1
neat-python==0.92
packaging==21.3
pluggy==1.0.0
py==1.11.0
pygame==2.1.2
pyparsing==3.0.7
pytest==7.1.0
tomli==2.0.1
